Add files via upload
barony (DLC enabled).exe is the most up to date version as of 4/2/22. Download it, rename it to barony.exe, and replace the version that is in your steam folder. Also delete a file called scores, as well as any save files. These files may crash the game as it boots up.